Moved stuff into separate file.
