+++ spark-orientdb-connector/src/main/scala/com/metreta/spark/orientdb/connector/ClassRDDFunctions.scala
@@ -3,46 +3,16 @@
 
 package com.metreta.spark.orientdb.connector
 
-import org.apache.spark.SparkContext
-import org.apache.spark.rdd.RDD
-import com.metreta.spark.orientdb.connector.api.OrientDBConnector
-import com.orientechnologies.orient.core.sql.OCommandSQL
-import org.apache.spark.Logging
-import java.text.SimpleDateFormat
-import com.orientechnologies.orient.core.metadata.schema.OClass
-import com.orientechnologies.orient.core.metadata.schema.OType
-import java.lang.reflect.Field
-import com.orientechnologies.orient.core.metadata.schema.OProperty
-import org.apache.commons.codec.binary.Base64
 import java.math.BigDecimal
+import java.text.SimpleDateFormat
+
+import com.metreta.spark.orientdb.connector.api.OrientDBConnector
+import com.orientechnologies.orient.core.metadata.schema.OType
 import com.orientechnologies.orient.core.record.impl.ODocument
-import org.apache.spark.TaskContext
-
-class ClassJsonRDDFunctions(rdd: RDD[String]) extends Serializable with Logging {
-  /**
-    * Saves an instance of [[org.apache.spark.rdd.RDD RDD]] into an OrientDB class.
-    * @param myClass the OrientDB class to save in
-    */
-  /*
-	 * Prerequisites:
-	 *  -> Input class must have been created on OrientDB
-	 *  -> rdd must be composed of strings in Json format
-	 */ def saveJsonToOrient(myClass: String)(implicit connector: OrientDBConnector = OrientDBConnector(rdd.sparkContext.getConf)): Unit = {
-    rdd.foreachPartition { partition ⇒
-      val db = connector.databaseDocumentTx()
-
-      while (partition.hasNext) {
-        val obj = partition.next()
-        val doc = new ODocument(myClass);
-        doc.fromJSON(obj)
-        db.save(doc)
-
-      }
-      db.commit()
-      db.close()
-    }
-  }
-}
+import com.orientechnologies.orient.core.sql.OCommandSQL
+import org.apache.commons.codec.binary.Base64
+import org.apache.spark.Logging
+import org.apache.spark.rdd.RDD
 
 class ClassRDDFunctions[T](rdd: RDD[T]) extends Serializable with Logging {
   /**
